dom/src/threads/test/test_functionHandlers.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads
-->
<head>
  <title>Test for DOM Worker Threads Recursion</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("functionHandlers_worker.js");

  var message = "hello!";

  worker.onmessage = function(event) {
    is(event.data, "uncaught exception: " + message, "Bad message!");
    SimpleTest.finish();
  };

  worker.postMessage(message);

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
